+++ lib/clients/copyFactory/streaming/transactionListener.py
@@ -14,3 +14,11 @@ class TransactionListener:
             transaction_event: Transaction event with an array of packets.
         """
         pass
+
+    async def on_error(self, error: Exception):
+        """Calls a predefined function with the received error.
+
+        Args:
+            error: Error received during retrieve attempt.
+        """
+        pass